% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dastardly_dois.R
\name{get_dastardly_dois}
\alias{get_dastardly_dois}
\title{Get a list of dastardly dois to cause problems in the doistext functions}
\usage{
get_dastardly_dois(limit, type = "ScholarlyArticle", keywords = "bird")
}
\arguments{
\item{limit}{A numerical value representing how many dois to return. This number will be larger than those returned because the function selects only the dois that lead to scientific articles}

\item{type}{A string of length 1 indicating the type of document from which to return dois}
}
\value{
A character vector of dois
}
\description{
Using datacite website to extract dois that link to scientific publications (can we extend to reports?)
}
