% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html.R
\name{parse_html}
\alias{parse_html}
\title{Convert html to plain text}
\usage{
parse_html(htm)
}
\arguments{
\item{htm}{A large string of length 1 containing the html for a journal article}
}
\value{
A character vector containing the plain text version of the input html document with paragraphs in separate lines.
}
\description{
Removes html from a downloaded article by calling the htm2txt package.
}
